// src/types/Supplier/Supplier.ts
export interface Supplier {
  Id: number;
  DocumentNumber: string;
  CompanyName: string;
  Address: string;
  Number: string;
  Neighborhood: string;
  City: string;
  State: string;
  PostalCode: string;
  Phone: string;
  RegistrationStatus: string;
  BranchType: string;
  Email: string;
  CreatedAt: string;
  UpdatedAt: string;
}